Distribution Finalisation
Pages are now evenly distributed from a set of approved positions which can be easily added to as the level expands in size. The events on each of the pages are randomised and once the player can move them, they will be able to piece together the histories of the different Archivists.